LayoutTests/fast/dom/css-mediarule-functions-expected.txt

This tests the insertRule(rule, index) and deleteRule(index) methods of the CSSMediaRule interface. It has passed if all of the output below begins with the text "PASS".

PASS: No exception raised! New rule inserted successfully.
PASS: Exception raised successfully. Type: IndexSizeError (DOM Exception 1): The index is not in the allowed range.
PASS: Exception raised successfully. Type: IndexSizeError (DOM Exception 1): The index is not in the allowed range.
PASS: Exception raised successfully. Type: SyntaxError (DOM Exception 12): The string did not match the expected pattern.
PASS: Exception raised successfully. Type: HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree.
PASS: Exception raised successfully. Type: SyntaxError (DOM Exception 12): The string did not match the expected pattern.
PASS: Exception raised successfully. Type: SyntaxError (DOM Exception 12): The string did not match the expected pattern.
PASS: No exception raised! Rule at position 'length - 1' deleted successfully.
PASS: Exception raised successfully. Type: IndexSizeError (DOM Exception 1): The index is not in the allowed range.
PASS: Exception raised successfully. Type: IndexSizeError (DOM Exception 1): The index is not in the allowed range.
